emit types from JS
remove the current (manually written) types, and replace them
with typings that are autogenerated from the JS code into
the `dist` folder (add a `prepare` script so that npm publish
will automatically run it).

for those types that were manually defined (mainly the "Plain"
interfaces) copy the code to `index.d.ts` because porting them
to JSDoc syntax would be too tedious.